requirements.txt
fpdf2==2.7.8
markdown==3.5.1
beautifulsoup4==4.12.2 